refresh the  template of the g8 template

+++ src/main/resources/g8-build-template/test.sh
@@ -2,8 +2,14 @@
 
 if [[ -f ./build.sbt ]] && [[ -d ./src/main/g8 ]]; then
 
+    if ! command -v g8 &> /dev/null
+    then
+        echo "[ERROR] g8 command cannot be found, please install g8 following http://www.foundweekends.org/giter8/setup.html"
+        exit -1
+    fi
+
     export TEMPLATE=`pwd | xargs basename`
-    echo ${TEMPLATE}
+    echo "Processing ${TEMPLATE} ..."
     mkdir -p $testTargetFolder$
     cd $testTargetFolder$
     sudo rm -r $testTemplateName$
@@ -12,9 +18,12 @@ if [[ -f ./build.sbt ]] && [[ -d ./src/main/g8 ]]; then
     $beforeTest$
     $testCommand$
 
+    echo "Done."
+    exit 0
+
 else
 
-    echo "WARNING: run the script ./test.sh in the template root folder"
+    echo "[ERROR] run the script ./test.sh in the template's root folder"
     exit -1
 
 fi
